Enne bug fixe ja eroor handlingu lisamist

+++ 03_21.02_klassikomplekt/kymnev6istlus/src/main/java/ee/tormi/kymnev6istlus/entity/Results.java
@@ -28,5 +28,5 @@ public class Results {
     private int javelin;
     private int thousandFiveHundredMeterRun;
 
-    private float totalScore;
+    private int totalScore;
 }
